2010-07-06  Atsushi Enomoto  <atsushi@ximian.com>
	* JsonSerializationWriter.cs : it cannot serialize DateTime in
	  double, which causes crash in deserializer.


svn path=/trunk/mcs/; revision=159945

+++ mcs/class/System.ServiceModel.Web/System.Runtime.Serialization.Json/JsonSerializationWriter.cs
@@ -101,7 +101,7 @@ namespace System.Runtime.Serialization.Json
 					writer.WriteString ("false");
 				break;
 			case TypeCode.DateTime:
-				writer.WriteString (String.Format (CultureInfo.InvariantCulture, "/Date({0})/", ((DateTime) graph).Subtract (new DateTime (1970, 1, 1)).TotalMilliseconds));
+				writer.WriteString (String.Format (CultureInfo.InvariantCulture, "/Date({0})/", (long) ((DateTime) graph).Subtract (new DateTime (1970, 1, 1)).TotalMilliseconds));
 				break;
 			default:
 				if (graph is Guid) {
